requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
crispy-bootstrap4==2024.10
Django==5.1.4
django-allauth==65.3.1
django-crispy-forms==2.3
django-cron==0.6.0
idna==3.10
pillow==11.0.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
stripe==11.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
